Delegate prices loading to library
