+++ src/lib/Functions.ts
@@ -83,3 +83,40 @@ export function enumKeys<T>(enumType: any): T[] {
     isNaN(Number(k))
   ) as unknown as T[];
 }
+
+export interface ExecParams {
+  context?: any;
+  runnable: (arg0: any) => any;
+  interval?: number;
+  attempts?: number;
+}
+
+export const INTERRUPT = `INTERRUPT`;
+export const SERVICE_LIMIT = `Service invoked too many times`;
+
+export function execute({
+  context,
+  runnable,
+  interval = 500,
+  attempts = 5,
+}: ExecParams): any {
+  let err: Error | any;
+  do {
+    try {
+      err = null;
+      return runnable(context);
+    } catch (e: any) {
+      err = e;
+      if (e.message.includes(INTERRUPT) || e.message.includes(SERVICE_LIMIT)) {
+        break;
+      }
+    }
+    if (attempts > 0) {
+      Utilities.sleep(interval);
+    }
+  } while (--attempts > 0);
+
+  if (err) {
+    throw err;
+  }
+}
